requirements.txt
flask
PyJWT
requests
cryptography
PyYAML
